homework-7 Done
